Embedding enhanced client within current client
